+++ src/main/java/racingcar/configuration/AppConfig.java
@@ -1,6 +1,6 @@
 package racingcar.configuration;
 
-import racingcar.controller.RacingControllerImpl;
+import racingcar.controller.RacingController;
 import racingcar.repository.MemoryRepository;
 import racingcar.repository.Repository;
 import racingcar.service.RacingService;
@@ -32,8 +32,8 @@ public class AppConfig {
         return new RacingServiceImpl(getRepository());
     }
 
-    public RacingControllerImpl getCarRacingController() {
-        return new RacingControllerImpl(getInputView(), getOutputView(),
+    public RacingController getCarRacingController() {
+        return new RacingController(getInputView(), getOutputView(),
                 getCarRacingService());
     }
 }
